chore(auth): move jwt util functions into own file for testing
